mitglied entfernen

+++ kleiderkammer/kleidung/templates/js/kleidung-details.js
@@ -103,19 +103,21 @@ $(function () {
 
         const kleidung_id = $("#kleidung-details .modal").data("kleidung-id");
 
-        $.ajax({
-            cache: false,
-            method: "DELETE",
-            success: function () {
-                // Entferne Zeile aus Tabelle und schließe Details
-                $(`.kleidung[data-id=${kleidung_id}]`).remove();
-                $("#kleidung-details").find(".modal").modal("hide");
-            },
-            url: "{{ url_for('kleidung_api.archivieren', kleidung_id='_kleidung_id_') }}".replace(
-                "_kleidung_id_",
-                kleidung_id
-            ),
-        });
+        if (confirm("Wirklich löschen?")) {
+            $.ajax({
+                cache: false,
+                method: "DELETE",
+                success: function () {
+                    // Entferne Zeile aus Tabelle und schließe Details
+                    $(`.kleidung[data-id=${kleidung_id}]`).remove();
+                    $("#kleidung-details").find(".modal").modal("hide");
+                },
+                url: "{{ url_for('kleidung_api.archivieren', kleidung_id='_kleidung_id_') }}".replace(
+                    "_kleidung_id_",
+                    kleidung_id
+                ),
+            });
+        }
     });
 });
